Parallel Grid issue fixes
